requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
cfgv==3.4.0
DateTime==5.4
distlib==0.3.8
Django==5.0
django-debug-toolbar==4.2.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
identify==2.5.33
nodeenv==1.8.0
platformdirs==4.1.0
pre-commit==3.6.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
virtualenv==20.25.0
zope.interface==6.1
